+++ src/rdkafka_queue.c
@@ -130,6 +130,7 @@ void rd_kafka_q_fwd_set0(rd_kafka_q_t *srcq,
                          rd_kafka_q_t *destq,
                          int do_lock,
                          int fwd_app) {
+        rd_assert(srcq != destq);
 
         if (do_lock)
                 mtx_lock(&srcq->rkq_lock);